In terminate volume check if exists.
In case a volume is deleted in the lightbits cluster and than
detach/deleted through openstack - operation will fail, volume will stay
"detaching".

We check if the volume exists before trying to change the acl - if its
return "Not Found" just return since there is no volume and no acl to
change.

Change-Id: If9cbaf6c73cad9e02a266f3038b64bff37fca00d